+++ packages/appcd-util/src/util.js
@@ -336,6 +336,52 @@ export function inherits(subject, base) {
 }
 
 /**
+ * Removes non-serializable properties and circular references from a value such that it can be
+ * printed, sent over an IPC channel, or JSON stringified.
+ *
+ * @param {*} it - The value to scrub.
+ * @returns {*}
+ */
+export function makeSerializable(it) {
+	let values = new Set();
+
+	try {
+		return (function dupe (src) {
+			const type = typeof src;
+
+			if (type === 'number' && isNaN(src)) {
+				return null;
+			}
+
+			if (src === null || type === 'string' || type === 'number' || type === 'boolean') {
+				return src;
+			}
+
+			if (type === 'object') {
+				if (values.has(src)) {
+					return;
+				}
+
+				values.add(src);
+
+				if (Array.isArray(src)) {
+					return src.map(dupe);
+				}
+
+				const obj = {};
+				for (let [ key, value ] of Object.entries(src)) {
+					obj[key] = dupe(value);
+				}
+				return obj;
+			}
+		}(it));
+	} finally {
+		values.clear();
+		values = null;
+	}
+}
+
+/**
  * Deeply merges two JavaScript objects.
  *
  * @param {Object} dest - The object to copy the source into.