+++ packages/appcd-core/CHANGELOG.md
@@ -11,6 +11,7 @@
    [(DAEMON-311)](https://jira.appcelerator.org/browse/DAEMON-311)
  * feat(server): Stop config service when daemon is gracefully shutdown.
  * feat(server): Added import for `android.buildTools.selectedVersion` Titanium config setting.
+ * feat(status): Added machine id to status info.
  * refactor: Updated to `appcd-config@3.0.0`, created the Appc Daemon config schema, and converted
    the `default.js` config to `default.json`.
  * refactor: Removed Titanium CLI Genymotion config import.